Creates model for storing URLs
This change adds a mongoose schema definition that lays out the format
for storing the URLs.  The first thing the schema tracks is the original
url which will be passed in the new action.  The second thing in the
schema is the short url.  This will be the token that is generated by
the app and returned to the user.  It will be a random 6 character
string ala lobste.rs.