Adds in application logic
This is a big change.  There are a few things that happen:

* The routes are wired to the controller
* The logic to create a new url is created
* The logic to decode a shorturl is created

The route change is fairly straightforward.  It adds two routes, one
that looks for `/new/*` and sends that for encoding and another that
looks for `/*` and sends that for decoding.

The logic for creating a new url took a little while for me to get
right.  I was originally going to just have an auto-incrementing field
for the `shortUrl`.  I realized that wasn't a good way when I had to
fight with Mongo to do it.  I came up with another plan which is used by
[lobste.rs](https://github.com/jcs/lobsters/blob/master/app/models/short_id.rb)
to generate a short random string of 6 characters.  I set a unique index
in the db so that should work for quite a while.

One of the other requirements was that is only allows valid looking URLs
so I found a package that provides that called `valid-url`.

To decode the shortUrl was pretty simple in comparison.  It just looks
in the db and returns what it finds or an error.